02python_db_usecase/requirements.txt
psycopg2-binary
python-dotenv
streamlit
invoice_generator